requirements.txt
streamlit>=1.32.0
python-chess>=1.999
google-genai>=0.8.0
requests>=2.31.0
reportlab>=4.0.0
